+++ stories/user-action/select/MultiSelect.stories.tsx
@@ -13,6 +13,7 @@ type Story = StoryObj<typeof meta>;
 export const multiSelect: Story = {
   args: {
     disabled: false,
+    invalid: false,
     onValuesChanged: action('onChange'),
     values: ['Apple', 'Cherry'],
     children: [